Moved the DataSet and SandboxDataSet back to admin service since it's the only service accessing the sandboxDataset properties. need to make a note for sandbox creation & testing to add sandbox-properties property tag  - a list of Sandbox datasets is needed.
# Only here for local Dev for Sandbox creation, not real dataSets
sandbox-properties:
  sandboxDatasets:
    - label: Demo Dataset
      id: demo-dataset
    - label: SBAC Dataset
      id: sbac-dataset
