+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/ApplicationConfiguration.java
@@ -3,6 +3,7 @@ package org.opentestsystem.rdw.admin;
 import org.opentestsystem.rdw.admin.client.DefaultConfigServerClient;
 import org.opentestsystem.rdw.admin.client.TenantConfigurationLookupClient;
 import org.opentestsystem.rdw.admin.client.TenantConfigurationLookupProperties;
+import org.opentestsystem.rdw.admin.multitenant.SandboxDataSets;
 import org.opentestsystem.rdw.admin.multitenant.TenantAdministrationConfiguration;
 import org.opentestsystem.rdw.admin.multitenant.TenantAdministrationProcessor;
 import org.opentestsystem.rdw.admin.service.impl.DefaultActuatorClientRdwService;
@@ -10,7 +11,6 @@ import org.opentestsystem.rdw.admin.service.impl.DefaultActuatorClientRdwService
 import org.opentestsystem.rdw.multitenant.messaging.TenantChannelInterceptor;
 import org.opentestsystem.rdw.multitenant.messaging.TenantMessageRequestResolver;
 import org.opentestsystem.rdw.reporting.common.multitenant.ReportingTenantMessageRequestResolver;
-import org.opentestsystem.rdw.reporting.common.multitenant.SandboxDataSets;
 import org.opentestsystem.rdw.reporting.common.stream.DefaultMessageSecurityService;
 import org.opentestsystem.rdw.reporting.common.stream.MessageSecurityService;
 import org.opentestsystem.rdw.reporting.common.web.security.jwt.JwtService;
@@ -178,8 +178,7 @@ public class ApplicationConfiguration {
     }
 
     @Bean
-    @ConfigurationProperties("sandbox-datasets")
-    @RefreshScope
+    @ConfigurationProperties("sandbox-properties")
     public SandboxDataSets sandboxDatasets() {
         return new SandboxDataSets();
     }
