+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/SandboxConfigurationPackage.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.admin.model;
 
+import org.opentestsystem.rdw.admin.multitenant.DataSet;
 import org.opentestsystem.rdw.multitenant.Tenant;
-import org.opentestsystem.rdw.reporting.common.multitenant.DataSet;
 
 import java.util.Set;
