+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/SandboxConfigurationStubs.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.admin.model;
 
 import java.util.HashSet;
 import java.util.Set;
-import org.opentestsystem.rdw.reporting.common.multitenant.DataSet;
+import org.opentestsystem.rdw.admin.multitenant.DataSet;
 
 
 /**
