+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/TenantConfiguration.java
@@ -2,9 +2,10 @@ package org.opentestsystem.rdw.admin.model;
 
 import com.fasterxml.jackson.annotation.JsonRawValue;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+
+import org.opentestsystem.rdw.admin.multitenant.SandboxDataSets;
 import org.opentestsystem.rdw.admin.web.RawJsonDeserializer;
 import org.opentestsystem.rdw.multitenant.Tenant;
-import org.opentestsystem.rdw.reporting.common.multitenant.SandboxDataSets;
 
 /**
  * This contains tenant specific configuration, tenant, properties and localization for a tenant.
