+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/multitenant/DataSet.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.common.multitenant;
+package org.opentestsystem.rdw.admin.multitenant;
 
 /**
  * DataSet is a preconfigured set of metadata the Sandbox uses.
