+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/multitenant/SandboxDataSets.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.common.multitenant;
+package org.opentestsystem.rdw.admin.multitenant;
 
 import java.util.List;
