+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/service/impl/DefaultSandboxService.java
@@ -5,12 +5,12 @@ import org.opentestsystem.rdw.admin.model.ApplicationSandboxConfigurations;
 import org.opentestsystem.rdw.admin.model.SandboxConfiguration;
 import org.opentestsystem.rdw.admin.model.SandboxConfigurationPackage;
 import org.opentestsystem.rdw.admin.model.TenantConfigurationStubs;
+import org.opentestsystem.rdw.admin.multitenant.SandboxDataSets;
 import org.opentestsystem.rdw.admin.repository.impl.InMemorySandboxConfigurationRepository;
 import org.opentestsystem.rdw.admin.service.SandboxService;
 import org.opentestsystem.rdw.multitenant.Tenant;
 import org.opentestsystem.rdw.multitenant.TenantProperties;
 import org.opentestsystem.rdw.reporting.common.multitenant.ReportingTenantIdResolver;
-import org.opentestsystem.rdw.reporting.common.multitenant.SandboxDataSets;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
