+++ admin-service/src/test/java/org/opentestsystem/rdw/admin/multitenant/SandboxDataSetsTest.java
@@ -1,6 +1,5 @@
-package org.opentestsystem.rdw.reporting.common.multitenant;
+package org.opentestsystem.rdw.admin.multitenant;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,13 +11,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-// TODO RP-394 Move some integration test (IT) to End to End tests (ETE) and configure to run in CI
-//this test only runs against a local docker instance
-// set this property to run this test (or temporarily comment out this annotation)
-// java -Dtenant-client.docker-it-enable=true
-@Ignore
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @ActiveProfiles({"sandbox_datasets"})
@@ -28,7 +23,6 @@ public class SandboxDataSetsTest {
     @Autowired
     SandboxDataSets sandboxDataSets;
 
-    @Ignore
     @Test
     public void itShouldLoadSandboxDataSets() {
         assertThat(sandboxDataSets).isNotNull();
@@ -36,7 +30,6 @@ public class SandboxDataSetsTest {
         assertThat(sandboxDataSets.getSandboxDatasets().size()).isEqualTo(2);
     }
 
-    @Ignore
     @Configuration
     static class Config {
