+++ admin-service/src/test/java/org/opentestsystem/rdw/admin/service/impl/DefaultSandboxServiceTest.java
@@ -8,12 +8,12 @@ import org.opentestsystem.rdw.admin.model.ApplicationSandboxConfigurations;
 import org.opentestsystem.rdw.admin.model.SandboxConfiguration;
 import org.opentestsystem.rdw.admin.model.SandboxConfigurationStubs;
 import org.opentestsystem.rdw.admin.model.TenantConfigurationStubs;
+import org.opentestsystem.rdw.admin.multitenant.SandboxDataSets;
 import org.opentestsystem.rdw.admin.repository.impl.InMemorySandboxConfigurationRepository;
 import org.opentestsystem.rdw.multitenant.TenantProperties;
 import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemProperties;
 import org.opentestsystem.rdw.reporting.common.multitenant.ReportingTenantIdResolver;
-import org.opentestsystem.rdw.reporting.common.multitenant.SandboxDataSets;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -72,6 +72,7 @@ public class DefaultSandboxServiceTest {
         assertThat(applicationSandboxConfigurations.getSandboxConfigurationPackage()).isNotNull();
         assertThat(applicationSandboxConfigurations.getSandboxes()).isNotEmpty();
         assertThat(applicationSandboxConfigurations.getSandboxes().size()).isEqualTo(2);
+        assertThat(applicationSandboxConfigurations.getSandboxConfigurationPackage().getDataSets().size()).isEqualTo(SandboxConfigurationStubs.getDataSets().size());
 
         Optional<SandboxConfiguration> sandboxCa = getSandboxConfigurationFromKey(applicationSandboxConfigurations, sandboxConfigurationCa.getSandbox().getKey());
         assertThat(sandboxCa).isPresent();
