--- common/src/test/resources/application-sandbox_datasets.yml
@@ -1,10 +1,0 @@
-logging:
-  level:
-    org.opentestsystem.rdw.admin: debug
-
-sandbox-properties:
-  sandboxDatasets:
-    - label: Demo Dataset
-      id: demo-dataset
-    - label: SBAC Dataset
-      id: sbac-dataset
